MB-39292: 3/n Integrate SharedMetaDataTable
The Collections::Manager (1 per bucket) is given 2 SharedMetaDataTable
objects, one for storing collection metadata and the second for storing
scope metadata.

The VB::ManifestEntry now moves two existing members over to the
SharedMetaDataTable, scope and maxTTL. The class is given a
SingleThreadedRCPtr reference to the metadata which is owned by
the Manager class. The metadata is further increased so that the
collection name is now readable using ManifestEntry::getName().

Scopes are changed so that there is no longer a set of known IDs, but a
map of ID to metadata, where the metadata stored is currently just the
scope's name.

With this change the scope and collections now have their name in-memory
enabling the next change, which will be the detection of a name change.

These changes take the VB::ManifestEntry from 88 bytes down to 72 bytes
which is a nice reduction for deployments with many collections. However
the scope set is now a map, so deployments with many scopes will see
a small increase in memory overheads.

Change-Id: I4fa52c7b1a063e4726f102efc2381dd592942b3d
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/144493
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>