MB-43546: Fix race between FollyExecutorPool::cancel & snooze
There is a potential race in FollyExecutorPool between the cancel()
and snooze() methods.

Unlike CB3ExecutorPool, cancel() isn't a completely blocking operation
- while the task _is_ marked as cancelled synchronously; the actual
deletion of the TaskProxy object is done lazily in the background
after cancel() returns.

This requires that other task operations (e.g. wake, snooze) check the
GlobalTask is not null before oeprating on it. This is correctly
handled for wake (see TaskProxy::wake), but _not_ for cancel. As a
result a snooze() which executes before the background cleanup of
taskOwners can result in a nullptr dereference.

Change-Id: I8ef476426b1056bfff646551fa38d94f4ff64f01
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/144166
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>