MB-43700: Generate the correct by-id range for collection 0xff
Code in the backfill_by_id which generates the range of keys for
the KVStore was calculating the end key incorrectly, this shows up
when collection 0xff is requested.

The code generated two leb128 prefixes, start and end, but the bug
is that end of the range was calculated as leb128(collection + 1),
generating the wrong value for how the by-id range works.

E.g.

* start leb128(0xff) = 0xff.01
* end   0xff + 1 = leb128(0x100) = 0x80.02

With the above range couchstore then returns 0 keys for the OSO
backfill because the start/end do not encompass any keys (end
appears less than start).

What should of happened is that the end is calculated from the start
value.

* start = leb128(0xff) = 0xff.01
* end   = start + 1    = 0xff.02

All 0xff keys will on disk have 0xff.01 as their 'leading' key bytes,
so a range of 0xff.01 to 0xff.02 will return all 0xff.01 prefixed keys.

Change-Id: I97e5b0b7932c68ca555d8eb3fc36ab46fca69ea3
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/144278
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>