MB-43478: [BP] Relax RR checks in the ReplicaEvictedBeforeActive test
The test is very sensitive to minor differences of how much memory the
ItemPager frees up in different runs, which makes the test failing
sometimes.

But, the original purpose of the test is just ensuring that Replica VBs
are ejected before Active VBs, so just verify that.

Change-Id: Idedec7485e616392caf32663ec546397115cffbe
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/144355
Reviewed-by: Dave Rigby <daver@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>