MB-43847: [Ephe] Backfill skips stale item at RangeItr::begin
Before this fix a Producer may send duplicate items in the Backfill
snapshot.
That breaks the definition of "snapshot" as a "sequence of unique keys",
so DCP clients may identify that as an unexpected condition.
In the specific case of DCP Consumer, replica is resilient to duplicates
for Normal mutations but it may fail on SyncWrites.

Change-Id: Ifb41e818270c411c62ccf6ae6ad959dee2631fe9
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/144261
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>