MB-32732: doStatsTask: Account deleted tasks to correct engine
ExecutorPool::doStatsTask() takes a reference to all tasks being
printed (to avoid holding a lock for the entire function
execution). If the task's refcount drops to 1 during this function
(i.e. because it is cancelled by some other thread), then
doStatsTask() will end up deleting that task.

If this happens then the memory will be freed (and accounted) against
no bucket; resulting in an accounting mismatch (mem_used will be too
high).

Fix by ensuring the task shared_ptr in doStatsTask is reset under a
memory guard for the bucket which owns it.

Change-Id: I503c378d67c032b5ef1de4b1e6792ce3735e18ac
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/144377
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>