MB-20640: move phosphor trace config to memcached.json
Allow for setting the phosphor configuration in memcached.json
via a tag named phosphor_config which use the format specified
for TraceConfig.

The default configuration looks like:

    buffer-mode:ring;buffer-size:20971520;enabled-categories:*

Note that if one tries to change the configuration via settings
you would need to stop and start phosphor tracing via IOCTL
commands in order for it to take effect.

Change-Id: I8f9ecde1eac53437e6f96b4a25d65e09380d5b0e
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/144480
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>