MB-43815: Hold vb stateLock while dropping prepare from DM
The durability monitor is guarded by the vb stateLock, ensure it is
locked (read/shared) when removing a prepare from the DM while dropping
a collection. This prevents a state change replacing the DM (e.g., a
passive DM with an active DM) while it is in use.

Dropping prepares from the DM is expected to be infrequent, and the
number of prepares in flight at any one time does have a practical
upper bound, so the impact of acquiring the stateLock for each dropped
prepare should be minimal.

Change-Id: I5f50c49e97e992e715ac6c2cc7463fee5a3da5fe
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/144616
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>