MB-38109: Simplify logic for generating phosphor trace
This patch simplifies generation of a trace dump by
adding the logic to format the trace dump to the
"begin()" call which would "block" until the executor
had formatted the trace dump (in another thread).

Also remove the "chunked read" method to fetch the trace
dump to a single get call as the old chunk retrieval
required the server to keep state between each call to
the get (and that the trace dump could only be read
out _once_ by a single client. So if that client crashed
on the client side the entire trace would be lost).

By removal of all the complexity on the chunked reads
with multiple locks between various objects we got
rid of the code with caused the crash in the
referenced MB.

Change-Id: Id2096147033c668bc105fd28a5cce34006f198ee
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/144622
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>