MB-43815: Make isKeyLogicallyDeleted a const check
For the principle of "least surprise", move call to dropKey out of
EphemeralVBucket::isKeyLogicallyDeleted(...).

Using it is now slightly more verbose (an `if` and a separate call to
dropKey(...)), but now isKeyLogicallyDeleted is a simple test, and
won't trip callers up by dropping the key if it _is_ logically deleted.

Change-Id: I129e252d72a1228676caae4b42e75b285db4cde9
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/144730
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>