MB-41658: Don't clear dirty flag unconditionally during warmup
There exists a race-condition between warmup, commit of prepared
SyncWrites and persistence which can result in the persistence
callback finding a non-dirty item. This throws an assert failure and
terminates memcached.

The issue is caused when the following sequence of events occur, on
various concurrent threads:

  0. Outstanding prepare exists on disk (without Commit).

  1. [Reader] Warmup loads prepare from disk.

  2. [NonIO] ActiveDM completes prepare, adding Committed SyncWrite to
     HashTable (as dirty) and queuing Committed SyncWrite to CkptMgr.

  3. [Writer] Flusher runs, writing the Commit to disk - but crucially
      *doesn't* get as far as executing PersistenceCallback.

  4. [Reader] Warmup advances to value load phase, and loads the
     Commit from [3] into HashTable (setting it clean).

  5. [Writer] Persistence Callback runs, encounters a clean item in HT
     for item just persisted - CRASH.

The underlying bug is at [4] - when loading items into the HashTable
during warmup; they should _not_ be unconditionally set as clean.

Bulk of the patch is the testcase setting up the above sequence of
events, fix itself is straightforard :)

Change-Id: I70b410848500bb2c5457020d747beb3f1637337b
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/144740
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>