[BP] MB-43744: Ensure we persist unclean shutdown failover state
Currently we may not persist the new FailoverTable entry that we
generate after an unclean shutdown if we do not do any mutations. If
we then do a clean shutdown we won't generate a new FailoverTable
entry and we may not roll back consumers that we should. Correct this
by queueing the setVBucketState after we add the vBucket to the vbMap.

Change-Id: I9913f0ba987f52fe39b3701a6317f3f1522a70b5
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/145779
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>