MB-44002: Move manipulation of session counter to core [1/2]
The use of the session counter (and its cas) originates to when the
core didn't know anything about the underlying engines types. With
all of the refactoring we've done with the engines (squashing all
of the repos so that all bucket types commands are known to the core)
it is easier to perform the session check in the core and have it
clean up the data as part of the command lifecycle.

The second patch in this series will move the handling of these
commands out of unknown_command (which is already pretty big)
and into separate functions in the engine interface.

Change-Id: Ife2e6e7405f163e5c00c2b853ef2f9b1aa3469a6
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/144813
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>