MB-43978: Fix intermittent failure in TaskStats memory tests
The following two patches added new tests for memory tracking for
ExecutorPool classes:

* MB-33900: Expand ExecutorPool memory tracking tests (3d768cc7b)
* MB-32732: doStatsTask: Account deleted tasks to correct engine (efc8e69c9)

To check memory usage after cancel() they polled, waiting for the
number of registered Tasks to be zero before reading memory usage.

However, this approach is flawed for CB3ExecutorPool, as it removes
Tasks from it's set of registered tasks before deleting them. As a
consequence, the test could intermittently fail, as the actual Task
had not _yet_ been deleted (and memory freed) by the background
thread.

Instead, simply poll for memory usage to equal the expected
value. Given this approach no longer needs to read the number of
tasks, remove the (test-only) code added to expose that information.

Change-Id: I35eccf6f76ad9583f7d46e8085134e1688818647
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/145194
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>