MB-44026: Remove and relax "MANIFEST_AHEAD" error
The manifest ahead error in some situations leads to a permanent loss of
replication, where a replica is ahead and the active will never catch
up. Instead of trying to come up with another recovery mechanism, relax
(and remove) the error and allow the 'future' vbucket to perform a
stream-request and just use rollback as we normally would.

Change-Id: Id22faa0976fff674a0fccf82a8ef0982fc7eccd1
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/145082
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>