Add command line utility to test authentication
Couchbase SASL supports a number of different mechanisms
to authenticate to the server, and not all mechanisms
work for all users (for instance if the user is defined
in LDAP only PLAIN will work).

In addition to that it is ns_server which provides
the passsword database to memcached where it have a
uniqe hashed version for each of the authentication
mechanism and memcached will try to perform the
same hash for the input data and compare the hashed
version with the data provided by ns_server.

ex:

$ ./mctestauth --user myuser --password foobar
SCRAM-SHA512: OK
SCRAM-SHA256: OK
SCRAM-SHA1: OK
PLAIN: OK

For an external user the output would look like:

./mctestauth --user myuser --password foobar
SCRAM-SHA512: FAILED - Authentication failed: Auth failure (32)
SCRAM-SHA256: FAILED - Authentication failed: Auth failure (32)
SCRAM-SHA1: FAILED - Authentication failed: Auth failure (32)
PLAIN: OK

Change-Id: Ic945a9b0960b1403b825441f98fe4d848b926fa1
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/145276
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>