MB-44102: Rollback compaction completion callback if it fails
The compaction completion callback is external to the KVStore layer
and could throw an exception. Doing so would cause the compaction
to fail and we should not update various cached stats in this case
as they won't reflect the on disk state after we "abort" the
compaction. Additionally we should move the completionCallback
invocation before the update of the other cached stats in the KVStore.

Change-Id: I98523339f4c6fa0fda5255b02bb85b984e0e7c64
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/145206
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>