MB-44098: Remove the 'new' file if we detect an error after creating it
The final phase of compaction moves the 'x.couch.y.compact' file to
be 'x.couch.y' and then the current revision is updated. Once the
current revision is updated and the vbLock dropped, flush now
writes to the new file.

However if an exception occurs after x.couch.y is created and
before the current revision is updated, x.couch.y remains on disk.

This could be fine if the vbucket (x) just keeps on moving forward,
however if x is deleted and created again, the new version of the
vbucket uses x.couch.y, which mixes the old and new vbucket together.

This is what finally led to the collection manifest decode function
later aborting, the vbucket merged old/new to create a new manifest
which actually just meant every collection appeared twice.

The fix in this commit is to enclose the final phase of compaction
in a try/catch block. The move of the file and finalisation code
is all moved into the try/catch block (inside the new
switchToNewFile function). A subtly of this is that the try/catch
is within the vbLock 'locked' scope, so that any exception that
is thrown is caught and handled with the vbLock held - this means
we are safe to 'rm' x.couch.y and cannot race with vbucket creation
trying to create and write to x.couch.y.

A secondary patch will be created which improves manifest write
path so it would abort before writing a broken manifest to disk.

Change-Id: I4365a0cc20d6fe2409d5c3d249603a8e69ce267d
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/145278
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>