MB-44079: Dump full ADM at State::updateHighCompletedSeqno()
Extra logging has been recently added for dumping the ADM::State in case
of error, but the resolvedQueue is an ADM member. Add the resolveQueue
to the dump.
Note that currently we print only the queue size and highEnqueuedSeqno
as the underlying type (folly::USPSCQueue) public API exposes only
enqueue/dequeue.

Change-Id: I09aa3eacd8732e37297f673f731628a410b0bff1
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/145386
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>