MB-44173: Initialize debug symbols on windows during startup
We need to initialize the symbol support on windows with
SymInitialize before looking up the symbols. Previously
we did that in our signal handler, but it looks like
that is "too late" and we're failing to load all of
the modules at that stage.

Given that we don't load additional libraries with LoadLibraries
we can just as well initialize the symbol tables during
startup.

Change-Id: I83d0f609f0a9c2bda319d2ceb51bc94f55531695
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/145416
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>