MB-44200: Reduce WARNING messages from endStream
DCP end stream is logging every 'end' other than 'OK' as a
WARNING. The day to day operations of a couchbase cluster will
see streams ending with reasons other than 'OK'. For example
a node-add and rebalance can see streams change state and it
is not an indication of a problem.

This commit categorises the DCP end reasons so that we reduce
the number of warnings logged.

Change-Id: Icabc83924da7543632ef18f963ce251e0d9bcf4d
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/145543
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@couchbase.com>