MB-44200: 1/2 When 'DcpStream' privilege is lost, end before snapshot
Before this change on loss of the DcpStream privilege, the stream is
ends, but this is done by pushing a stream-end-response. This means
the stream ends after sending a snapshot.

This can be improved to reduce 'data' being seen by switching the
snapshot marker for an end-stream and processing no further messages
from the stream.

An issue this raises though is that in some circumstances, the
ActiveStream is set to dead, but remains owned by the DcpProducer.
If the snapshot was large, then the dead stream will have the snapshot
items in it's readyQueue - patch 2/2 will address that.

Change-Id: I040ea2b03443e1f81a4e4c756a8b044efb35cece
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/145544
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>