MB-43818: Add kv_symbolize_backtrace.py
Script to convert addresses from backtraces printed via
print_backtrace() to symbol names.

Accepts input from stdin or files provided on the command line; for
each line checks to see if it matches the backtrace format and if so
invokes gdb to symbolise the address.

If the line doesn't match then simply passes that line through - as
such this can be used as a filter on a complete log.

Also accepts --source-line flag to print the source file and line
details. This is omitted by default as it makes a very verbose output.

Example usage:

    cat memcached.log | grep CRITICAL | ./kv_symbolize_backtrace.py

Change-Id: I68d839e23703a402dd93fd99686854f14c53d3b0
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/145545
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>