MB-44139: [Ephe] Correctly update num-deleted-item in deleteStoredValue
In EphemeralVBucket::softDeleteStoredValue we try to update an existing
OSV and move it to the end of the SeqList if possible.
That may be prevented by range-reads that cover that OSV. In that case
we add a new stored value at the end of the SeqList and leave the old
OSV at its place.

Before this patch we used to always update the SeqList::numDeletedItems
counter based on the semantic of an "update SV", also for when the "add
SV" logic is executed.

That led to missing to increment the num of deletions in the case where
a Replica vbucket receives multiple deletions in a row for the same key
(which is legal for a Replica due to deduplication at master).

As a side effect, the non-negative num-deleted-items counter may
underflow in the HTCleaner as soon as tombstones are purged.

Change-Id: I5c95926441b1a676628d77e4dc1ff6f56b167652
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/145711
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>