MB-44200: 2/2 Destruct dead ActiveStreams when they're finished
In the case when the producer has the control
"send_stream_end_on_client_close_stream" set to false (the default
value) and a stream ends for any reason other than a client
initiated close (e.g. stream reaches the end-seqno), the stream
becomes 'dead' and will remain in the producer's map. It would
only destruct if a second stream was created (replacing the
dead one) or if the producer was destructed (connection closed).

This patch changes this so that once a dead stream has reached the
'end-message' it will be removed from the map, and resources will
be freed.

Change-Id: I600db4f298020313e9962e9301c2419bd585165e
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/145667
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>