MB-44167: Notify durable op clients when collection dropped
When we drop a collection we remove the elements in the DM belonging
to that collection. Currently we don't notify the clients which can
hang a shutdown. Notify the clients when we remove a SyncWrite from
the ADM. The PDM shouldn't need notifications as nothing should be
waiting for a replica to commit something.

Change-Id: I84a6af998eeb7d3c6adab5d58817d7b187fe4f41
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/145922
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>