MB-42224: Add test for sync-header failures at KVStore::snapshotVBucket
Current opinion is that we have nothing to fix in the
KVStore::snapshotVBucket code path with regard to sync-header failures.

The test shows that the major side-effect of a sync-failure is that the
new vbstate is visible if the user makes a call to
KVStore::getPersistedVBucketState(), while the cached vbstate stays
correctly untouched.

Given that KVStore::getPersistedVBucketState() is a test-only API and
that all prod code in KV uses KVStore::getCachedVBucketState(), then we
are fine.

Change-Id: I9180ad61a83e0081c8ac957ac1a1f001469af615
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/146203
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>