MB-42224: Remove KVStore::setVBucketState()
The semantic of the function is the same of the existing
KVStore::updateCachedVBState(). While the latter is a protected function
called within KVStore::snapshotVBucket, setVBucketState() is a public
one and expected to be called by the user after a successful
KVStore::commit().

Remove setVBucketState() and update the cached vbstate within
KVStore::commit.

Change-Id: I3ca752523e0b2f17192af8f334f39d2998566370
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/146117
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>