MB-44390: Do not consider age of ineligible values in eviction threshold
The PagingVisitor tracks the age of visited items in a histogram. The
visitor uses this to determine a threshold below which a certain
percentage of items should fall. This is used to protect a percentage
of "young" items from being evicted immediately; their MFU might not
reflect how hot they are yet.

However, the visitor includes the age of non-resident items in the
histogram. Thus, rather than protecting X% of _resident_ items, it
protects X% of _all_ items.

If a bucket is 30% resident, and eviction protects the youngest 30% of
items, it's possible _no_ resident item would be considered old enough
to evict.

With this patch, the visitor _does not_ add the age of non-resident
items to the histogram at all. Now the age threshold will be determined
from the distribution of MFUs of resident items only.

Change-Id: If5626d1cd9cc925f704d0ed15da5475e432286ef
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/147600
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>