MB-44390: Do not bias eviction MFU histogram for ineligible values
The PagingVisitor tracks the MFU value of visited items in a histogram.
The visitor uses this to determine an MFU threshold below which a
certain percentage of items should fall.

Previously, if an item were selected for eviction but was later found
to be ineligible (dirty, non-resident, or a prepare) the maximum MFU
value would be added instead of the real MFU value.

This was done to bias the histogram "upward"; the intuition being that
the visitor failed to evict an item, so the threshold may need to
slightly increase to ensure the required percentage of items are
evicted.

However, for value eviction buckets this functionally treated every
non-resident item as extremely "hot".

The PagingVisitor seeks to evict the X% coldest the _resident_ values,
however because non-resident items are still included in the histogram
count, the percentile returned from the histogram becomes the Xth
percentile MFU of _all_ items.

Simplified example:

 10 resident items
 90 non-resident items
 visitor trying to evict the 10% coldest resident items

The intent would be to evict 1 item.

Instead, the visitor would evict all 10 resident items, as they are the
10% coldest of _all_ items, because the non-resident items are treated
as having max MFU.

This patch now _does not_ add non-resident items' MFU to the histogram
at all. Now the MFU threshold will be determined from the distribution
of MFUs of resident items only.

Change-Id: Id00aa6b3592e99e4cee59c4e8a5beecd63b333db
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/146372
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>