MB-44417: Stop concurrent dereference calls throwing
The SharedMetaDataTable::dereference function doesn't account for the
real usage, in that multiple threads can concurrently dereference the
same meta, so only the SharedMetaDataTable is the owner (refCount of 1).
When this happens, all of those threads will still call dereference, the
first thread will clean-up the map and all subsequent calls will throw
an exception.

The approach here is to keep the exception, but change the API so that
the caller must give back their "RCPtr" (by use of move/rvalue). The
dereference function is now where the RCPtr::reset occurs.

Change-Id: Ic885fb44e13d55b3f267e31302dc481eb2448157
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/146451
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>