MB-44414: Actually invalidate chains when moving writes to resolvedQ
Currently when we move ActiveSyncWrites to the resolvedQueue we call
ActiveSyncWrite::initializeChains(nullptr, nullptr) which actually
does nothing as the passed chains are nullptrs. We should be setting
the chains to nullptr as a transition to active with a different
chain and then from active to non-active would cause us to have
invalid pointers in ActiveSyncWrite objects in trackedWrites. Should
we then attempt to dump the state of this DurabilityMonitor we'd
segfault dereferencing the invalid ptr.

Correct this by actually resetting the chains.

Change-Id: Id076db7b560a1754dd68caa81fc54d5cf7cb7154
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/146563
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>