Optimize send buffer use
Calculate the total number of bytes we want to send for an
individual response and and ask bufferevent to allocate the
space (in an iovector) at the same time.

Previously we would use a number of calls to bufferevent_write
to write the individual parts of a packet and each of these
calls would then result in (potentially) allocating new buffers.
To work around the problem for the headers we would first copy
everything into the front end threads scratch buffer, before
using bufferevent_write with the header extras and key and then
a second call with the value.

We may use bufferevent a bit more optimal by calculating the
total amount of data we're going to send and then request (at
at least) this amount of memory to become available in an
iovector. The first entry of the iovector will be "the current
tail chunk" and the second entry is a new buffer which is big
enough to hold the remaining data.

Change-Id: Ifb8b49e65d60de1ca943fbe65211f28094a6fc0c
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/146916
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>