MB-44558: HashTable: Reset all item counts on clear()
Not all members of Statistics are reset to zero on clear:

* numDeletedItems - used to calculate curr_items stat amongst others.
* numSystemItems - used to calculate curr_items stat amongst others.
* numPreparedSyncWrites - used to calculate curr_items stat amongst
                          others.
* metaDataMemory - used by ItemPager to calculate pagable memory.

Part of the reason this occurred was that the stats struct
contained members of type std::atomic, which doesn't allow
assignment. As such, the reset() method of Statistics couldn't use the
simpler implementation:

    stats = {};

but instead had to explicitly assign each to zero:

    stats.foo.store(0);
    stats.bar.store(0);
    ...

As such, this made it easy to forget to reset 'stats.baz' when it was added.

Fix by introducing a helper class - CopyableAtomic - which wraps
std::atomic but allows copying (and assignment). By changing the type
of each stat to this template type, we can use the simpler and less
error-prone method to reset.

Change-Id: I549d0fb8c7bab483f1a0b4d2a7734844a7172cc5
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/146924
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>