MB-41257: Abort warmup if we fail to make KVFileHandle to vbucket file
During Warmup::loadCollectionStatsForShard() we try and collect
collection's stats from disk. As a prerequisite of calling
KVStore::getCollectionStats() we need to pass it a ref to a KVFileHandle
that will allow us to access KVStore for the current vbucket. If
KVStore::makeFileHandle() returns a nullptr this was cause us to de-ref
an invalid pointer.

To prevent this check the output of KVStore::makeFileHandle(), if its
nullptr then abort warmup as we're unable to access the data file on
disk. Which will cause problems further down the line in warmup.

Change-Id: Ieb6ba4c801ff16f9f294d6c5f4e6ecc88c6a4d43
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/146961
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>