MB-41257: Abort warmup if we're unable to scan the KVStore for prepares
Currently if EPBucket::loadPreparedSyncWrites() fails to read from disk
during Warmup::loadPreparedSyncWrites(), we won't be aware that this has
occurred and treat it as if there were no prepares on disk. Which could
leave our durability monitors in an inconsistent state. This also
indicates we might not be able to reliably access the data files on
disk, if we're unable to access them now just after reading them to
create KVStore objects.

So to make Warmup::loadPreparedSyncWrites() aware that a failure to read
data files has occurred, add a success bool to struct
KVBucketIface::LoadPreparedSyncWritesResult{}. If this is set to false
then abort warmup so we don't warmup into a bad state.

loadPreparedSyncWrites() is also called during rollback, if it returns
success as false we will abort the rollback task as we where unable to
read from the on disk data file.

This patch also adds a unit test for the warmup case.

Change-Id: I8df4666464a33962c09f9bd6fdea5f77c4f50a8e
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/147058
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>