MB-44452: Relax memory ordering of RCValue refcount
RCValue (the base class used for items managed by SingleThreadedRCPtr)
contains a reference count as an atomic <int>. Currently we use the
default memory ordering (memory_order_seq_cst). This is overly
pessimistic - the increments of refcounts do not need such a strict
memory order, and the decrements only need acquire-release.

Abstract the actual reference count into it's own 'RefCount' class,
which exposes just the operations required (with appropriate memory
ordering).

Also adjust HashTableBench/Clear so the HashTable is the only owner of
each Blob, so we also benchmark Blob delete cost.

Note that benchmarking on mancouch (x86-64) doesn't show any
measurable impact in this change, and indeed examining the assembly of
compiled code shows the same instructions (lock add/sub) are
generated. However for other architectures (e.g. AArch64) this should
be an optimisation, and even on x86064 the compiler _should_ have more
leeway to reorder instructions when the STRCPtr methods are inlined.

Change-Id: I6287468a1236e0d9405d25e143d161ec362bcfc1
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/147069
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>