MB-44452: Make StoredValue::ordered const & non-atomic
The StoredValue::ordered flag specifies what "subclass" a storedvalue
is - StoredValue or OrderedStoredValue. It is currently located within
the AtomicBitSet array of flags.

This is sub-optimal, because AtomicBitSet:

- Forces the entire word to be read atomically every time the bit is
  accessed, and

- Prevents the compiler from re-ordering loads to 'ordered'.

Given the flag is const and hence will never change, we can move it to
the "non-atomic" set of flags, which allows more efficient code to be
generated.

For example, consider the (partial) assembly generated for
StoredValue::Deleter::operator():

    void StoredValue::Deleter::operator()(StoredValue* val) {
        if (val->isOrdered()) {
            delete static_cast<OrderedStoredValue*>(val);
        } else {
            delete val;
        }
    }

Extract from AtomicBitSet implementation (current) which is selecting
between deleting as OrderedStoredValue or StoredValue; plus inlined
destructor:

               ...
  (A1)         movzbl 0x33(%rsi),%eax
               test   $0x8,%al
             ↓ je     b0
               test   %rsi,%rsi
             ↓ je     92
  (A2)         movzbl 0x33(%rsi),%eax
               test   %al,%al
             ↓ js     110
         35:   mov    %rbx,%rdi
             → callq  ObjectRegistry::onDeleteStoredValue

Note the byte load of AtomicBitSet occurs mutliple times (A1, A2, more
not shown...); for every time the code needs to check a flag in that
byte; even the same flag.

Compare this to moving the flag so no longer atomic - it it checked
once (via testb) and then doesn't need to be re-checked (the compiler
has restructured the asm significantly so not seening exactly the same
asm):

               ....
  (A1)         testb  $0x1,0x34(%rsi)
             ↓ jne    60
               ...
         60:   movzbl 0x33(%rsi),%eax
               test   $0x8,%al
             ↓ je     70
               movq   $0x0,0x8(%rsi)
         70:   mov    %rbx,%rdi
               movabs $0xffffffffffff,%rbp
             → callq  ObjectRegistry::onDeleteStoredValue
               ...

Results in 4.7% reduction in runtime of HashTableBench/Clear:

Before:

    -------------------------------------------------------------------------------------
    Benchmark                                           Time             CPU   Iterations
    -------------------------------------------------------------------------------------
    HashTableBench/Clear/iterations:100_median   17809656 ns     17809053 ns            5

After (non-atomic flag):

    -------------------------------------------------------------------------------------
    Benchmark                                           Time             CPU   Iterations
    -------------------------------------------------------------------------------------
    HashTableBench/Clear/iterations:100_median   16965649 ns     16964898 ns            5

Change-Id: Ie1a521b9f9dd7ee915caafa07ba29e9d5d019baa
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/147202
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>